refactor: split Flame region-map and next-turn hosts (#4582)
Move camera, warp markers, hover, turn-resolution session apply, and
idle-civilian rows into sibling libraries so wave-20 AC5 files sit at
or under 240 lines without changing map or confirmation behavior.
